add build flag for privval building
