+++ internal/state/tx_filter_test.go
@@ -23,8 +23,8 @@ func TestTxFilter(t *testing.T) {
 		tx    types.Tx
 		isErr bool
 	}{
-		{types.Tx(cmtrand.Bytes(2155)), false},
-		{types.Tx(cmtrand.Bytes(2156)), true},
+		{types.Tx(cmtrand.Bytes(2123)), false},
+		{types.Tx(cmtrand.Bytes(2124)), true},
 		{types.Tx(cmtrand.Bytes(3000)), true},
 	}
