+++ types/block_test.go
@@ -463,10 +463,10 @@ func TestBlockMaxDataBytes(t *testing.T) {
 		0: {-10, 1, 0, true, 0},
 		1: {10, 1, 0, true, 0},
 		2: {841, 1, 0, true, 0},
-		3: {842, 1, 0, false, 0},
-		4: {843, 1, 0, false, 1},
-		5: {954, 2, 0, false, 1},
-		6: {1053, 2, 100, false, 0},
+		3: {874, 1, 0, false, 0},
+		4: {875, 1, 0, false, 1},
+		5: {1018, 2, 0, false, 1},
+		6: {1117, 2, 100, false, 0},
 	}
 
 	for i, tc := range testCases {
@@ -494,8 +494,8 @@ func TestBlockMaxDataBytesNoEvidence(t *testing.T) {
 		0: {-10, 1, true, 0},
 		1: {10, 1, true, 0},
 		2: {841, 1, true, 0},
-		3: {842, 1, false, 0},
-		4: {843, 1, false, 1},
+		3: {874, 1, false, 0},
+		4: {875, 1, false, 1},
 	}
 
 	for i, tc := range testCases {
